requirements.txt
#   not complete versions missing !!!!
opentelemetry-api
opentelemetry-sdk
opentelemetry-instrumentation-logging
opentelemetry-exporter-otlp
# opentelemetry-exporter-jaeger-thrift
opentelemetry-instrumentation-fastapi
fastapi
uvicorn
